+++ src/Stages/StageLabor.java
@@ -1,9 +1,7 @@
 package Stages;
 
-import java.awt.Graphics;
-
-import DBZ.Main;
 import Images.GameImages;
+import java.awt.Graphics;
 
 public class StageLabor extends Stage {
 
@@ -14,7 +12,7 @@ public class StageLabor extends Stage {
 	
 	public void paint(Graphics g)
 	{	
-		g.drawImage(GameImages.maps[12], 0,0,(int)(1000),(int)(600),null);		
+		g.drawImage(GameImages.maps[12], 0,0, 1000, 600,null);
 	}
