+++ src/Stages/StageLookout.java
@@ -1,12 +1,9 @@
 package Stages;
 
 
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Graphics;
-
-import DBZ.Main;
 import Images.GameImages;
+import java.awt.Color;
+import java.awt.Graphics;
 
 
 public class StageLookout extends Stage{
@@ -26,7 +23,7 @@ public class StageLookout extends Stage{
 	public void paint(Graphics g)
 	{
 		g.setColor(new Color(150,150,250));
-		g.fillRect( 0,0,(int)(1000),(int)(600));
+		g.fillRect( 0,0, 1000, 600);
 	   
 		clouds.paintBackgroundClouds(g);
